Refactor: Move useProjects hook to hooks directory and update imports
